+++ src/main/java/csc1035/project2/RoomHandler.java
@@ -1,10 +1,13 @@
 package csc1035.project2;
 
 import java.awt.print.Book;
+import java.sql.Time;
+import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 
 // This might end up mapping to the Reservations table. See how we use it.
+// TODO: Please someone switch the room number to integers and remove all instances of `Integer.parseInt`
 public class RoomHandler {
 
 
@@ -38,6 +41,12 @@ public class RoomHandler {
         return bookings;
     }
 
+    /*
+    public void reserveRoomStudent ( Students s, Room r, Timestamp st, Timestamp et ) {
+        Booking b = new Booking( Integer.parseInt(r.getNum()), st, et );
+    }
+    */
+
     /**
      * Method to return all rooms with a reservation.
      *
@@ -46,7 +55,6 @@ public class RoomHandler {
     public List<Room> getReservedRooms () {
         List<Room> tmp = new ArrayList<>();
         for ( Room r : this.getRooms() ) {
-            // TODO: Please someone switch the room number to integers!
             for ( Booking b : this.getBookings() ) {
                 if ( Integer.parseInt(r.getNum()) == b.getNum() ) {
                     tmp.add( r );
@@ -56,13 +64,22 @@ public class RoomHandler {
         return tmp;
     }
 
-
-    public List<Room> getAvailableRooms () {
-        List<Room> tmp = new ArrayList<>();
-        List<Room> reserveRooms = new ArrayList<>(this.getReservedRooms());
-        for ( Room r : this.getRooms() ) {
-            if ( !reserveRooms.contains( r ) ) {
-                tmp.add( r );
+    /**
+     * Method to return all rooms available at a specified time.
+     *
+     * <p>Method checks if a room is within a booking and if the timestamp given is within the allocated time of
+     * the booking. If so then it removes it from the temporary list and returns the remaining rooms.</p>
+     *
+     * @param t Timestamp to check against
+     * @return List of all available rooms at a specified time.
+     */
+    public List<Room> getAvailableRooms ( Timestamp t ) {
+        List<Room> tmp = new ArrayList<>( this.getRooms() );
+        for ( Room r : tmp ) {
+            for ( Booking b : this.getBookings() ) {
+                if ( Integer.parseInt(r.getNum()) == b.getNum() && (t.after(b.getStart()) && t.before(b.getEnd())) ) {
+                    tmp.remove( r );
+                }
             }
         }
         return tmp;