+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ##### v0.2 [03.03.2021]: Room Booking Methids
 * (ERICK) Created `getReservedRooms()` method.
-* (ERICK) Created 'getAvailableRooms()' method.
+* (ERICK) Created `getAvailableRooms()` method.
 * (ERICK) Booking ID is now auto-generated.
+* (ERICK) `getAvailableRooms()` now takes timestamp.